trade-executor/requirements.txt
temporalio==1.4.0
ccxt==4.2.15
asyncpg==0.29.0
python-dotenv==1.0.1
pydantic==2.6.1
fastapi==0.109.2
uvicorn==0.27.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-multipart==0.0.9
aiohttp==3.9.3
tenacity==8.2.3
prometheus-client==0.19.0
structlog==24.1.0
aiokafka==0.10.0
confluent-kafka==2.3.0
psycopg2-binary==2.9.9 